Travis message: Improve readability and message text
Dima remarked that saying our own docs are not official is sort of
weird. This was also inconsistent from the homepage which said nothing
of the sort.

Also made the message a bit more pleasant to read. Not sure if the
eyesore factor makes people read or not read but it is now gone.