script: fix push when multiple CI pipelines run at same time (#27)
If pushing does not work the first time, we now scrap the cloned target
repo, clone it again, and copy and commit the new content again, then
try another push. We do so twice, and after that we fail as before.

Bycatch: add comments for hard-coded parameters and more consistently
use "target repository" instead of "publishing repository".